LPS-1 Do not use 'api' in package path
