+++ modules/apps/frontend-editor/frontend-editor-alloyeditor-web/src/main/java/com/liferay/frontend/editor/alloyeditor/web/internal/constants/AlloyEditorConstants.java
@@ -26,7 +26,7 @@ public class AlloyEditorConstants {
 	 * <p>
 	 * Do not change this value unless you stop implementing the legacy {@link
 	 * com.liferay.portal.kernel.editor.Editor} interface and leave the {@link
-	 * com.liferay.frontend.editor.api.EditorRenderer} alone; otherwise, the
+	 * com.liferay.frontend.editor.EditorRenderer} alone; otherwise, the
 	 * former will fail because it hard codes the use of this namespace.
 	 * </p>
 	 */
