+++ modules/apps/frontend-editor/frontend-editor-alloyeditor-web/src/main/java/com/liferay/frontend/editor/alloyeditor/web/internal/editor/embed/YouTubeEditorEmbedProvider.java
@@ -14,8 +14,8 @@
 
 package com.liferay.frontend.editor.alloyeditor.web.internal.editor.embed;
 
-import com.liferay.frontend.editor.api.embed.EditorEmbedProvider;
-import com.liferay.frontend.editor.api.embed.EditorEmbedProviderTypeConstants;
+import com.liferay.frontend.editor.embed.EditorEmbedProvider;
+import com.liferay.frontend.editor.embed.EditorEmbedProviderTypeConstants;
 import com.liferay.petra.string.StringBundler;
 
 import org.osgi.service.component.annotations.Component;
