+++ modules/apps/frontend-editor/frontend-editor-ckeditor-web/src/main/java/com/liferay/frontend/editor/ckeditor/web/internal/CKEditorBBCodeEditor.java
@@ -14,7 +14,7 @@
 
 package com.liferay.frontend.editor.ckeditor.web.internal;
 
-import com.liferay.frontend.editor.api.EditorRenderer;
+import com.liferay.frontend.editor.EditorRenderer;
 import com.liferay.frontend.editor.ckeditor.web.internal.constants.CKEditorConstants;
 import com.liferay.portal.kernel.editor.Editor;
 import com.liferay.portal.kernel.servlet.PortalWebResourceConstants;
