+++ modules/apps/frontend-editor/frontend-editor-simple-web/src/main/java/com/liferay/frontend/editor/simple/web/internal/SimpleEditor.java
@@ -14,7 +14,7 @@
 
 package com.liferay.frontend.editor.simple.web.internal;
 
-import com.liferay.frontend.editor.api.EditorRenderer;
+import com.liferay.frontend.editor.EditorRenderer;
 import com.liferay.frontend.editor.simple.web.internal.constants.SimpleEditorConstants;
 import com.liferay.portal.kernel.editor.Editor;
 import com.liferay.portal.kernel.servlet.PortalWebResourceConstants;
