+++ modules/apps/frontend-editor/frontend-editor-taglib/src/main/java/com/liferay/frontend/editor/taglib/servlet/taglib/ResourcesTag.java
@@ -14,7 +14,7 @@
 
 package com.liferay.frontend.editor.taglib.servlet.taglib;
 
-import com.liferay.frontend.editor.api.EditorRenderer;
+import com.liferay.frontend.editor.EditorRenderer;
 import com.liferay.frontend.editor.taglib.internal.EditorRendererUtil;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.servlet.PortalWebResourcesUtil;
