+++ modules/apps/frontend-editor/frontend-editor-tinymce-web/src/main/java/com/liferay/frontend/editor/tinymce/web/internal/TinyMCESimpleEditor.java
@@ -14,7 +14,7 @@
 
 package com.liferay.frontend.editor.tinymce.web.internal;
 
-import com.liferay.frontend.editor.api.EditorRenderer;
+import com.liferay.frontend.editor.EditorRenderer;
 import com.liferay.frontend.editor.tinymce.web.internal.constants.TinyMCEEditorConstants;
 import com.liferay.portal.kernel.editor.Editor;
 import com.liferay.portal.kernel.servlet.PortalWebResourceConstants;
