+++ modules/apps/frontend-editor/frontend-editor-tinymce-web/src/main/java/com/liferay/frontend/editor/tinymce/web/internal/constants/TinyMCEEditorConstants.java
@@ -26,7 +26,7 @@ public class TinyMCEEditorConstants {
 	 * <p>
 	 * Do not change this value unless you stop implementing the legacy {@link
 	 * com.liferay.portal.kernel.editor.Editor} interface and leave the {@link
-	 * com.liferay.frontend.editor.api.EditorRenderer} alone; otherwise, the
+	 * com.liferay.frontend.editor.EditorRenderer} alone; otherwise, the
 	 * former will fail because it hard codes the use of this namespace.
 	 * </p>
 	 */
